Block states

Armon.SolverState — Type

SolverState

Object containing all non-constant parameters needed to run the solver, as well as type-parameters needed to avoid runtime dispatch.

This object is local to a block (or set of blocks): multiple blocks could be at different steps of the solver at once.

Armon.first_state — Function

first_state(grid::BlockGrid)

A SolverState which can be used as a global state when outside of a solver cycle. It belongs to the first device block.

Armon.SolverStep — Module

SolverStep

Enumeration of each state a LocalTaskBlock can be in. block_state_machine advances this state.

Armon.block_state_machine — Function

block_state_machine(params::ArmonParameters, blk::LocalTaskBlock)

Advances the SolverStep state of the blk, apply each step of the solver on the blk. This continues until the current cycle is done, or the block needs to wait for another block to do the ghost cells exchange (block_ghost_exchange) or compute the its time step (next_time_step).

Returns true if we reached the end of the current cycle for blk.

Time step reduction

Armon.next_time_step — Function

next_time_step(params::ArmonParameters, state::SolverState, blk::LocalTaskBlock; already_contributed=false)
next_time_step(params::ArmonParameters, state::SolverState, grid::BlockGrid)

Compute the time step of the next cycle. This is done at the start of the current cycle.

Since the current cycle does not rely on an up-to-date time step, the time step reduction is done fully asynchronously, including the global MPI reduction. The accuracy cost of this optimisation is minimal, as the CFL condition prevents the time step from being too large. Additionally, we prevent the time step from increasing of more than +5% of the previous one.

For first cycle, if no initial time step is given, the time step of the next cycle is reused for the initial cycle.

If blk is given, its contribution is only added to the state.global_dt (the GlobalTimeStep). Passing the whole block grid will block until the new time step is computed.

Armon.GlobalTimeStep — Type

GlobalTimeStep

Holds all information about the current time and time step for the current solver cycle. This struct is global and shared among all blocks.

When reaching next_time_step, blocks will contribute to the calculation of the next time step. The last block doing so will start the MPI reduction. The first block reaching the start of the next cycle will wait until this reduction is completed, updating the GlobalTimeStep when so.

Armon.TimeStepState.WaitingForMPI — Constant

One thread is waiting for the MPI reduction to complete

Armon.TimeStepState.Done — Constant

MPI is done: current_dt is up-to-date

Armon.TimeStepState.Ready — Constant

current_dt is up-to-date and blocks can contribute to next_dt

Armon.TimeStepState.DoingMPI — Constant

The MPI reduction has started

Armon.TimeStepState.AllContributed — Constant

All blocks contributed to next_dt, one thread will start the MPI reduction

Block exchanges

Armon.block_ghost_exchange — Function

block_ghost_exchange(params::ArmonParameters, state::SolverState, blk::LocalTaskBlock)

Handles communications between the blk neighbours, along the current state.axis. If blk is on one of the edges of the grid, a remote exchange is performed with the neighbouring RemoteTaskBlock, or the global boundary conditions are applied.

Returns true if exchanges were not completed, and the block is waiting on another to be ready for the exchange.

Armon.start_exchange — Function

start_exchange(
    params::ArmonParameters,
    blk::LocalTaskBlock{D, H}, other_blk::RemoteTaskBlock{B}, side::Side.T
) where {D, H, B}

Start the exchange between one local block and a remote block from another sub-domain. Returns true if the exchange is BlockExchangeState.Done, false if BlockExchangeState.InProgress.

Armon.finish_exchange — Function

finish_exchange(
    params::ArmonParameters,
    blk::LocalTaskBlock{D, H}, other_blk::RemoteTaskBlock{B}, side::Side.T
) where {D, H, B}

Finish the exchange between one local block and a remote block from another sub-domain, if MPI communications are done. Returns true if the exchange is BlockExchangeState.Done, false if BlockExchangeState.InProgress.

Armon.BlockInterface — Type

BlockInterface

Represents the interface between two neighbouring TaskBlocks. It synchronizes the state of the blocks to make sure the halo exchange happens when both blocks are ready, and that the operation is done by only one of the blocks.

Armon.mark_ready_for_exchange! — Function

mark_ready_for_exchange!(blk, side)

Mark blk in the interface along its side as ready for an exchange.

Return true if the blk should do the exchange, and the new [BlockExchangeState] of the interface.

Armon.exchange_done! — Function

exchange_done!(blk, side)

Mark the exchange of the interface of the blk along side as done. The other block need to acknowledge this before the exchange state can be reset.

Armon.side_is_done! — Function

side_is_done!(blk::LocalTaskBlock, side::Side.T, done::Bool=true)

Set the interface of blk along side as done. This is a non-atomic operation, meant only to avoid repeating the exchange logic multiple times.

Armon.is_side_done — Function

is_side_done(blk::LocalTaskBlock, side::Side.T)

The value set by side_is_done!.

Armon.BlockExchangeState — Module

BlockExchangeState

State of an interface between two blocks, controlling if a cell exchange can happen or did happen.

Armon.BlockExchangeState.NotReady — Constant

One of the blocks is not ready yet

Armon.BlockExchangeState.InProgress — Constant

One of the blocks is performing the exchange

Armon.BlockExchangeState.Done — Constant

The exchange is done