Blocking

Grid and blocks

Armon.BlockGrid — Type

BlockGrid{T, DeviceA, HostA, BufferA, Ghost, BlockSize, Device, SolverState}

Stores TaskBlocks on the Device and host memory, in a grid.

LocalTaskBlock are stored separately depending on if they have a StaticBSize of BlockSize (in blocks) or if they have a DynamicBSize (in edge_blocks).

Blocks have Ghost cells padding their real cells. This is included in their block_size. A block cannot have a number of real cells along an axis smaller than the number of ghost cells, unless there is only a single block in the grid along that axis.

"Edge blocks" are blocks located on the right and/or top edge of the grid. They exist in order to handle domains with dimensions which are not multiples of the block size.

DeviceA and HostA are AbstractArray types for the device and host respectively.

BufferA is the type of storage used for MPI buffers. MPI buffers are homogenous: they are either all on the host or all on the device.

Armon.grid_dimensions — Function

grid_dimensions(params::ArmonParameters)
grid_dimensions(block_size::NTuple{D, Int}, domain_size::NTuple{D, Int}, ghost::Int) where {D}

Returns the dimensions of the grid in the form (grid_size, static_sized_grid, remainder_block_size) from the block_size (the size of blocks in the static_sized_grid), the domain_size (number of real cells) and the number of ghost cells, common to all blocks.

grid_size is the static_sized_grid including the edge blocks. Edge blocks along the axis d have a size of remainder_block_size[d] along d only if they are at the edge of the grid along d, or block_size otherwise.

block_size includes the ghost cells in its dimensions, which must all be greater than 2*ghost.

If prod(block_size) == 0, then block_size is ignored and the grid is made of only a single block of size domain_size.

Note

Blocks must not be smaller than ghost, therefore edge blocks might be made bigger than block_size.

Note

In case domain_size is smaller than block_size .- 2*ghost along any axis, the grid will contain only edge blocks.

Armon.TaskBlock — Type

TaskBlock{V}

Abstract block used for cache blocking.

Armon.LocalTaskBlock — Type

LocalTaskBlock{D, H, Size, SolverState} <: TaskBlock{V}

Container of Size and variables of type D on the device and H on the host. Part of a BlockGrid.

The block stores its own solver state, allowing it to run all solver steps independantly of all other blocks, apart from steps requiring synchronization.

Armon.BlockData — Type

BlockData{V}

Holds the variables of all cells of a LocalTaskBlock.

Armon.RemoteTaskBlock — Type

RemoteTaskBlock{B} <: TaskBlock{B}

Block located at the border of a BlockGrid, containing MPI buffer of type B for communication with other BlockGrids.

Armon.device_to_host! — Function

device_to_host!(blk::LocalTaskBlock)

Copies the device data of blk to the host data. A no-op if the device is the host.

device_to_host!(grid::BlockGrid)

Copies device data of all blocks to the host data. A no-op if the device is the host.

Armon.host_to_device! — Function

device_to_host!(blk::LocalTaskBlock)

Copies the host data of blk to its device data. A no-op if the device is the host.

device_to_host!(grid::BlockGrid)

Copies host data of all blocks to the device data. A no-op if the device is the host.

Armon.buffers_on_device — Function

buffers_on_device(::BlockGrid)
buffers_on_device(::Type{<:BlockGrid})

true if the communication buffers are stored on the device, allowing direct transfers without passing through the host (GPU-aware communication).

Armon.device_is_host — Function

device_is_host(::BlockGrid{T, D, H})
device_is_host(::Type{<:BlockGrid{T, D, H}})

true if the device is the host, i.e. device blocks and host blocks are the same (and D == H).

Armon.all_blocks — Method

all_blocks(grid::BlockGrid)

Simple iterator over all blocks of the grid, excluding RemoteTaskBlocks.

Armon.block_idx — Function

block_idx(grid::BlockGrid, idx::CartesianIndex)

Linear index in grid.blocks of the block at idx in the statically-sized grid.

Armon.edge_block_idx — Function

edge_block_idx(grid::BlockGrid, idx::CartesianIndex)

Linear index in grid.edge_blocks of the block at idx, along the (dynamically-sized) edges of the grid.

Armon.remote_block_idx — Function

remote_block_idx(grid::BlockGrid, idx::CartesianIndex)

Linear index in grid.remote_blocks of the remote block at idx in the grid.

Armon.EdgeBlockRegions — Type

EdgeBlockRegions(grid::BlockGrid; real_sizes=false)
EdgeBlockRegions(
    grid_size::NTuple{D, Int}, static_sized_grid::NTuple{D, Int},
    block_size::NTuple{D, Int}, remainder_block_size::NTuple{D, Int}; ghosts=0
)

Iterator over edge blocks, their positions and their size in each edge region of the grid.

ghosts only affects the size of edge blocks: B .- 2*ghosts. If real_sizes == true, then ghosts is ghosts(grid), therefore the real block size of edge blocks is returned.

There is a maximum of 2^D-1 edge regions in a grid (see this explanation). When grid_size[i] != static_sized_grid[i], there cannot be an edge region, hence there are exactly 2^sum(grid_size .!= static_sized_grid)-1 regions.

julia> collect(Armon.EdgeBlockRegions((5, 5), (4, 4), (32, 32), (16, 16)))
3-element Vector{Tuple{Int64, CartesianIndices{2, Tuple{UnitRange{Int64}, UnitRange{Int64}}}, Tuple{Int64, Int64}}}:
 (4, CartesianIndices((5:5, 1:4)), (16, 32))
 (4, CartesianIndices((1:4, 5:5)), (32, 16))
 (1, CartesianIndices((5:5, 5:5)), (16, 16))

julia> collect(Armon.EdgeBlockRegions((5, 5, 5), (4, 5, 4), (32, 32, 32), (16, 16, 16)))
3-element Vector{Tuple{Int64, CartesianIndices{3, Tuple{UnitRange{Int64}, UnitRange{Int64}, UnitRange{Int64}}}, Tuple{Int64, Int64, Int64}}}:
 (20, CartesianIndices((5:5, 1:5, 1:4)), (16, 32, 32))
 (20, CartesianIndices((1:4, 1:5, 5:5)), (32, 32, 16))
 (5, CartesianIndices((5:5, 1:5, 5:5)), (16, 32, 16))

Armon.RemoteBlockRegions — Type

RemoteBlockRegions(grid::BlockGrid)
RemoteBlockRegions(grid_size::NTuple{D, Int})

Iterator of all remote block positions in each region (the faces of the grid). There is always 2*D regions.

julia> collect(Armon.RemoteBlockRegions((5, 3)))
4-element Vector{Tuple{Armon.Side.T, CartesianIndices{2, Tuple{UnitRange{Int64}, UnitRange{Int64}}}}}:
 (Armon.Side.Left, CartesianIndices((0:0, 1:3)))
 (Armon.Side.Right, CartesianIndices((6:6, 1:3)))
 (Armon.Side.Bottom, CartesianIndices((1:5, 0:0)))
 (Armon.Side.Top, CartesianIndices((1:5, 4:4)))

Armon.@iter_blocks — Macro

@iter_blocks for blk in all_blocks(grid)
    # body...
end

Applies the body of the for-loop in to all blocks of the grid.

The body is duplicated for inner and edge blocks, ensuring type-inference.

If params.use_multithreading, then an attempt will be made at equilibrating workload among threads.

# Iterate on all blocks of the grid
@iter_blocks for blk in all_blocks(grid)
    some_function(blk)
end

Armon.block_pos_containing_cell — Function

block_pos_containing_cell(grid::BlockGrid, pos::Union{CartesianIndex, NTuple})

Returns two CartesianIndexes: the first is the position of the block containing the cell at pos, the second is the position of the cell in that block.

Armon.block_origin — Function

block_origin(grid::BlockGrid, pos, include_ghosts=false)

A Tuple of the position of the cell at the bottom left corner of the LocalTaskBlock at pos in the grid.

If include_ghosts == true, then the cell position includes all ghost cells of the grid.

pos can be any of Integer, NTuple{N, Integer} or CartesianIndex.

Armon.block_at — Function

block_at(grid::BlockGrid, idx::CartesianIndex)

The TaskBlock at position idx in the grid.

Block size and iteration

Armon.BlockSize — Type

BlockSize

Dimensions of a LocalTaskBlock.

Armon.StaticBSize — Type

StaticBSize{S, Ghost} <: BlockSize

A BlockSize of size S and Ghost cells. S is embedded in the type, this reduces the amount of memory in the parameters of every kernel, as well as allowing the compiler to make some minor optizations in indexing expressions.

Ghost cells are included in S: there are S .- 2*Ghost real cells.

Armon.DynamicBSize — Type

DynamicBSize{Ghost} <: BlockSize

Similar to StaticBSize, but for blocks with a less-than-ideal size: block size is therefore not stored in the type. This results in less compilation when testing for different domain sizes with a constant StaticBSize.

The number of Ghost cells is still embedded in the type, as it can simplify some indexing expressions and for coherency.

Armon.block_size — Function

NTuple of the dimensions of a BlockSize

Armon.real_block_size — Function

NTuple of the dimensions of a BlockSize, excluding ghost cells

Armon.ghosts — Function

Number of ghost cells of a BlockSize

Armon.border_domain — Function

border_domain(bsize::BlockSize, side::Side.T; single_strip=true)

DomainRange of the real cells along side.

If single_strip == true, it includes only one "strip" of cells, that is length(border_domain(bsize, side)) == size_along(bsize, side). Otherwise, there are ghosts(bsize) strips of cells: all real cells which would be exchanged with another block along side.

Armon.ghost_domain — Function

ghost_domain(bsize::BlockSize, side::Side.T; single_strip=true)

DomainRange of all ghosts cells of side, excluding the corners of the block.

If single_strip == true, then the domain is only 1 cell thick, positionned at the furthest ghost cell from the real cells. Otherwise, there are ghosts(bsize) strips of cells: all ghost cells which would be exchanged with another block along side.

Armon.block_domain_range — Function

block_domain_range(bsize::BlockSize, corners)
block_domain_range(bsize::BlockSize, bottom_left::Tuple, top_right::Tuple)

A DomainRange built from offsets from the corners of bsize.

block_domain_range(bsize, (0, 0), (0, 0)) is the domain of all real cells in the block. block_domain_range(bsize, (-g, -g), (g, g)) would be the domain of all cells (real cells + g ghost cells) in the block.

Armon.position — Function

position(bsize::BlockSize, i)

N-dim position of the i-th cell in the block.

If 1 ≤ position(bsize, i)[d] ≤ block_size(bsize)[d] then the cell is not a ghost cell along the d dimension. See is_ghost.

Armon.lin_position — Function

lin_position(bsize::BlockSize, I)

From the NTuple (e.g. returned from position), return the linear index in the block. lin_position(bsize, position(bsize, i)) == i.

Armon.in_grid — Function

in_grid(idx, size)
in_grid(start, idx, size)

true if each axis of idx is between start and size. start defaults to 1.

Argument types can any mix of Integer, Tuple or CartesianIndex.

julia> in_grid(1, (1, 2), 2)  # same as `in_grid((1, 1), (1, 2), (2, 2))`
true

julia> in_grid((3, 1), (3, 2))
true

julia> in_grid((1, 3), (3, 2))
false

in_grid(idx, grid, axis::Axis.T)
in_grid(start, idx, grid, axis::Axis.T)

Same as in_grid(start, idx, grid), but only checks along axis.

Armon.is_ghost — Function

is_ghost(bsize::BlockSize, i, o=0)

true if the i-th cell of the block is a ghost cell, false otherwise.

o would be a "ring" index: o == 1 excludes the first ring of ghost cells, etc.

Armon.BlockRowIterator — Type

BlockRowIterator(grid::BlockGrid; kwargs...)
BlockRowIterator(grid::BlockGrid, blk::LocalTaskBlock; kwargs...)
BlockRowIterator(grid::BlockGrid, blk₁_pos, blk₂_pos; kwargs...)
BlockRowIterator(grid::BlockGrid, sub_domain::NTuple{2, CartesianIndex}; kwargs...)
BlockRowIterator(grid::BlockGrid, row_iter::CartesianIndices; global_ghosts=false, all_ghosts=false)

Iterate the rows of all blocks of the grid, row by row (and not block by block). This allows to iterate the cells of the grid as if it was a single block.

Elements are tuples of (block, global_row_idx, row_range). row_range is the range of cells in block for the current row.

Giving blk will return an iterator on the rows of the block.

Giving blk₁_pos and blk₂_pos will return an iterator over all rows between those blocks.

Giving sub_domain will return an iterator including only the cells contained in sub_domain. sub_domain is a cuboid defined by the position of the cells in the whole domain of grid, using its lower and upper corners.

row_iter is a iterator over global row indices.

If global_ghosts == true, then the ghost cells of at the border of the global domain are also returned. If all_ghosts == true, then the ghost cells of at the border of all blocks are also returned.

julia> params = ArmonParameters(; N=(24, 8), nghost=4, block_size=(20, 12), use_MPI=false);

julia> grid = BlockGrid(params);

julia> for (blk, row_idx, row_range) in Armon.BlockRowIterator(grid)
           println(Tuple(blk.pos), " - ", row_idx, " - ", row_range)
       end
(1, 1) - (1, 1) - 85:96
(2, 1) - (2, 1) - 85:96
(1, 1) - (1, 2) - 105:116
(2, 1) - (2, 2) - 105:116
(1, 1) - (1, 3) - 125:136
(2, 1) - (2, 3) - 125:136
(1, 1) - (1, 4) - 145:156
(2, 1) - (2, 4) - 145:156
(1, 2) - (1, 5) - 85:96
(2, 2) - (2, 5) - 85:96
(1, 2) - (1, 6) - 105:116
(2, 2) - (2, 6) - 105:116
(1, 2) - (1, 7) - 125:136
(2, 2) - (2, 7) - 125:136
(1, 2) - (1, 8) - 145:156
(2, 2) - (2, 8) - 145:156

Armon.DomainRange — Type

DomainRange

Two dimensional range to index a 2D array stored with contiguous rows. Not equivalent to CartesianIndices as it handles StepRanges properly.

Block Tree

Armon.BlockTree — Type

BlockTree(grid::BlockGrid, branches_per_level::Vector{Int})

A tree-like structure containing sub_blocks (other BlockTrees) as well as blocks (LocalTaskBlocks). It contains branches_per_level[1] branches, themselves containing branches_per_level[2] branches each, etc... The last level contains branches_per_level[end] LocalTaskBlocks, and no branches.

depth(bt) == 0 is the root node.

Technically, there is no topology constraint to which block may be part of the tree: for now blocks are naively assigned during the construction of the tree.

branches_per_level might specify a larger amount of blocks than there is in grid. In this case the extra blocks are not added to the tree, and empty branches are pruned.

Armon.all_blocks — Method

all_blocks(bt::BlockTree)

Iterator over every LocalTaskBlocks contained in bt (recursive).

Armon.is_leaf — Function

is_leaf(bt::BlockTree)

true if bt has no children, and contains only LocalTaskBlocks.

Armon.depth — Function

depth(bt::BlockTree)

Depth of bt in the parent tree.

Armon.tree_block_count — Function

tree_block_count(bt::BlockTree)

Number of LocalTaskBlocks contained in bt (recursive).

Armon.visit_all_tree_block — Function

visit_all_tree_block(f, bt::BlockTree; topdown=true, min_depth=0, max_depth=typemax(Int))

Recursively calls f(pos::Tuple, mb_child::BlockTree) for bt and its children.

If topdown == true, then f is called on parents before their children.

f is called only for BlockTrees with a depth ≥ min_depth and ≤ max_depth, defaulting to the whole BlockTree. Children with a depth > max_depth will not be recursed into.

Armon.iter_tree_block — Function

iter_tree_block(f, bt::BlockTree)

Calls f(blk::LocalTaskBlock) for all LocalTaskBlocks in bt.

Armon.apply_until_true — Function

apply_until_true(f, bt::BlockTree, sub_blocks_range=eachindex(bt.sub_blocks), repeats=2)

Applies recursively f on all LocalTaskBlocks of bt. As long as f returns false, it will be repeatedly called up to repeats^max_depth times for the block (max_depth being the depth of the LocalTaskBlock relative to bt).

Armon.block_levels_for_machine — Function

block_levels_for_machine(params::ArmonParameters; include_ghosts=false, kwargs...)
block_levels_for_machine(
    data_type::Type, block_count::Int, block_size::Tuple, vars_per_block::Int;
    levels=[:L3_cache, :L2_cache, :L1_cache]
)

Attempts to compute efficient sizes for branches of a BlockTree for the current machine using its topology and the cache sizes.

Block states

Armon.SolverState — Type

SolverState

Object containing all non-constant parameters needed to run the solver, as well as type-parameters needed to avoid runtime dispatch.

This object is local to a block (or set of blocks): multiple blocks could be at different steps of the solver at once.

Armon.first_state — Function

first_state(grid::BlockGrid)

A SolverState which can be used as a global state when outside of a solver cycle. It belongs to the first device block.

Armon.SolverStep — Module

SolverStep

Enumeration of each state a LocalTaskBlock can be in. block_state_machine advances this state.

Armon.block_state_machine — Function

block_state_machine(params::ArmonParameters, blk::LocalTaskBlock)

Advances the SolverStep state of the blk, apply each step of the solver on the blk. This continues until the current cycle is done, or the block needs to wait for another block to do the ghost cells exchange (block_ghost_exchange) or compute the its time step (next_time_step).

Returns true if we reached the end of the current cycle for blk.

Time step reduction

Armon.next_time_step — Function

next_time_step(params::ArmonParameters, state::SolverState, blk::LocalTaskBlock; already_contributed=false)
next_time_step(params::ArmonParameters, state::SolverState, grid::BlockGrid)

Compute the time step of the next cycle. This is done at the start of the current cycle.

Since the current cycle does not rely on an up-to-date time step, the time step reduction is done fully asynchronously, including the global MPI reduction. The accuracy cost of this optimisation is minimal, as the CFL condition prevents the time step from being too large. Additionally, we prevent the time step from increasing of more than +5% of the previous one.

For first cycle, if no initial time step is given, the time step of the next cycle is reused for the initial cycle.

If blk is given, its contribution is only added to the state.global_dt (the GlobalTimeStep). Passing the whole block grid will block until the new time step is computed.

Armon.GlobalTimeStep — Type

GlobalTimeStep

Holds all information about the current time and time step for the current solver cycle. This struct is global and shared among all blocks.

When reaching next_time_step, blocks will contribute to the calculation of the next time step. The last block doing so will start the MPI reduction. The first block reaching the start of the next cycle will wait until this reduction is completed, updating the GlobalTimeStep when so.

Armon.TimeStepState.WaitingForMPI — Constant

One thread is waiting for the MPI reduction to complete

Armon.TimeStepState.Done — Constant

MPI is done: current_dt is up-to-date

Armon.TimeStepState.Ready — Constant

current_dt is up-to-date and blocks can contribute to next_dt

Armon.TimeStepState.DoingMPI — Constant

The MPI reduction has started

Armon.TimeStepState.AllContributed — Constant

All blocks contributed to next_dt, one thread will start the MPI reduction

Block exchanges

Armon.block_ghost_exchange — Function

block_ghost_exchange(params::ArmonParameters, state::SolverState, blk::LocalTaskBlock)

Handles communications between the blk neighbours, along the current state.axis. If blk is on one of the edges of the grid, a remote exchange is performed with the neighbouring RemoteTaskBlock, or the global boundary conditions are applied.

Returns true if exchanges were not completed, and the block is waiting on another to be ready for the exchange.

Armon.start_exchange — Function

start_exchange(
    params::ArmonParameters,
    blk::LocalTaskBlock{D, H}, other_blk::RemoteTaskBlock{B}, side::Side.T
) where {D, H, B}

Start the exchange between one local block and a remote block from another sub-domain. Returns true if the exchange is BlockExchangeState.Done, false if BlockExchangeState.InProgress.

Armon.finish_exchange — Function

finish_exchange(
    params::ArmonParameters,
    blk::LocalTaskBlock{D, H}, other_blk::RemoteTaskBlock{B}, side::Side.T
) where {D, H, B}

Finish the exchange between one local block and a remote block from another sub-domain, if MPI communications are done. Returns true if the exchange is BlockExchangeState.Done, false if BlockExchangeState.InProgress.

Armon.BlockInterface — Type

BlockInterface

Represents the interface between two neighbouring TaskBlocks. It synchronizes the state of the blocks to make sure the halo exchange happens when both blocks are ready, and that the operation is done by only one of the blocks.

Armon.mark_ready_for_exchange! — Function

mark_ready_for_exchange!(blk, side)

Mark blk in the interface along its side as ready for an exchange.

Return true if the blk should do the exchange, and the new [BlockExchangeState] of the interface.

Armon.exchange_done! — Function

exchange_done!(blk, side)

Mark the exchange of the interface of the blk along side as done. The other block need to acknowledge this before the exchange state can be reset.

Armon.side_is_done! — Function

side_is_done!(blk::LocalTaskBlock, side::Side.T, done::Bool=true)

Set the interface of blk along side as done. This is a non-atomic operation, meant only to avoid repeating the exchange logic multiple times.

Armon.is_side_done — Function

is_side_done(blk::LocalTaskBlock, side::Side.T)

The value set by side_is_done!.

Armon.BlockExchangeState — Module

BlockExchangeState

State of an interface between two blocks, controlling if a cell exchange can happen or did happen.

Armon.BlockExchangeState.NotReady — Constant

One of the blocks is not ready yet

Armon.BlockExchangeState.InProgress — Constant

One of the blocks is performing the exchange

Armon.BlockExchangeState.Done — Constant

The exchange is done