Blocks

Armon.TaskBlock — Type

TaskBlock{V}

Abstract block used for cache blocking.

Armon.LocalTaskBlock — Type

LocalTaskBlock{D, H, Size, SolverState} <: TaskBlock{V}

Container of Size and variables of type D on the device and H on the host. Part of a BlockGrid.

The block stores its own solver state, allowing it to run all solver steps independantly of all other blocks, apart from steps requiring synchronization.

Armon.BlockData — Type

BlockData{V}

Holds the variables of all cells of a LocalTaskBlock.

Armon.RemoteTaskBlock — Type

RemoteTaskBlock{B} <: TaskBlock{B}

Block located at the border of a BlockGrid, containing MPI buffer of type B for communication with other BlockGrids.

Block size

Armon.BlockSize — Type

BlockSize

Dimensions of a LocalTaskBlock.

Armon.StaticBSize — Type

StaticBSize{S, Ghost} <: BlockSize

A BlockSize of size S and Ghost cells. S is embedded in the type, this reduces the amount of memory in the parameters of every kernel, as well as allowing the compiler to make some minor optizations in indexing expressions.

Ghost cells are included in S: there are S .- 2*Ghost real cells.

Armon.DynamicBSize — Type

DynamicBSize{Ghost} <: BlockSize

Similar to StaticBSize, but for blocks with a less-than-ideal size: block size is therefore not stored in the type. This results in less compilation when testing for different domain sizes with a constant StaticBSize.

The number of Ghost cells is still embedded in the type, as it can simplify some indexing expressions and for coherency.

Armon.block_size — Function

NTuple of the dimensions of a BlockSize

Armon.real_block_size — Function

NTuple of the dimensions of a BlockSize, excluding ghost cells

Armon.ghosts — Function

Number of ghost cells of a BlockSize

Block indexing

Armon.border_domain — Function

border_domain(bsize::BlockSize, side::Side.T; single_strip=true)

DomainRange of the real cells along side.

If single_strip == true, it includes only one "strip" of cells, that is length(border_domain(bsize, side)) == size_along(bsize, side). Otherwise, there are ghosts(bsize) strips of cells: all real cells which would be exchanged with another block along side.

Armon.ghost_domain — Function

ghost_domain(bsize::BlockSize, side::Side.T; single_strip=true)

DomainRange of all ghosts cells of side, excluding the corners of the block.

If single_strip == true, then the domain is only 1 cell thick, positionned at the furthest ghost cell from the real cells. Otherwise, there are ghosts(bsize) strips of cells: all ghost cells which would be exchanged with another block along side.

Armon.block_domain_range — Function

block_domain_range(bsize::BlockSize, corners)
block_domain_range(bsize::BlockSize, bottom_left::Tuple, top_right::Tuple)

A DomainRange built from offsets from the corners of bsize.

block_domain_range(bsize, (0, 0), (0, 0)) is the domain of all real cells in the block. block_domain_range(bsize, (-g, -g), (g, g)) would be the domain of all cells (real cells + g ghost cells) in the block.

Armon.position — Function

position(bsize::BlockSize, i)

N-dim position of the i-th cell in the block.

If 1 ≤ position(bsize, i)[d] ≤ block_size(bsize)[d] then the cell is not a ghost cell along the d dimension. See is_ghost.

Armon.lin_position — Function

lin_position(bsize::BlockSize, I)

From the NTuple (e.g. returned from position), return the linear index in the block. lin_position(bsize, position(bsize, i)) == i.

Armon.in_grid — Function

in_grid(idx, size)
in_grid(start, idx, size)

true if each axis of idx is between start and size. start defaults to 1.

Argument types can any mix of Integer, Tuple or CartesianIndex.

julia> in_grid(1, (1, 2), 2)  # same as `in_grid((1, 1), (1, 2), (2, 2))`
true

julia> in_grid((3, 1), (3, 2))
true

julia> in_grid((1, 3), (3, 2))
false

in_grid(idx, grid, axis::Axis.T)
in_grid(start, idx, grid, axis::Axis.T)

Same as in_grid(start, idx, grid), but only checks along axis.

Armon.is_ghost — Function

is_ghost(bsize::BlockSize, i, o=0)

true if the i-th cell of the block is a ghost cell, false otherwise.

o would be a "ring" index: o == 1 excludes the first ring of ghost cells, etc.

Armon.DomainRange — Type

DomainRange

Two dimensional range to index a 2D array stored with contiguous rows. Not equivalent to CartesianIndices as it handles StepRanges properly.