Grid

Armon.BlockGrid — Type

BlockGrid{T, DeviceA, HostA, BufferA, Ghost, BlockSize, Device, SolverState}

Stores TaskBlocks on the Device and host memory, in a grid.

LocalTaskBlock are stored separately depending on if they have a StaticBSize of BlockSize (in blocks) or if they have a DynamicBSize (in edge_blocks).

Blocks have Ghost cells padding their real cells. This is included in their block_size. A block cannot have a number of real cells along an axis smaller than the number of ghost cells, unless there is only a single block in the grid along that axis.

"Edge blocks" are blocks located on the right and/or top edge of the grid. They exist in order to handle domains with dimensions which are not multiples of the block size.

DeviceA and HostA are AbstractArray types for the device and host respectively.

BufferA is the type of storage used for MPI buffers. MPI buffers are homogenous: they are either all on the host or all on the device.

Armon.grid_dimensions — Function

grid_dimensions(params::ArmonParameters)
grid_dimensions(block_size::NTuple{D, Int}, domain_size::NTuple{D, Int}, ghost::Int) where {D}

Returns the dimensions of the grid in the form (grid_size, static_sized_grid, remainder_block_size) from the block_size (the size of blocks in the static_sized_grid), the domain_size (number of real cells) and the number of ghost cells, common to all blocks.

grid_size is the static_sized_grid including the edge blocks. Edge blocks along the axis d have a size of remainder_block_size[d] along d only if they are at the edge of the grid along d, or block_size otherwise.

block_size includes the ghost cells in its dimensions, which must all be greater than 2*ghost.

If prod(block_size) == 0, then block_size is ignored and the grid is made of only a single block of size domain_size.

Note

Blocks must not be smaller than ghost, therefore edge blocks might be made bigger than block_size.

Note

In case domain_size is smaller than block_size .- 2*ghost along any axis, the grid will contain only edge blocks.

Grid indexing

Armon.block_idx — Function

block_idx(grid::BlockGrid, idx::CartesianIndex)

Linear index in grid.blocks of the block at idx in the statically-sized grid.

Armon.edge_block_idx — Function

edge_block_idx(grid::BlockGrid, idx::CartesianIndex)

Linear index in grid.edge_blocks of the block at idx, along the (dynamically-sized) edges of the grid.

Armon.remote_block_idx — Function

remote_block_idx(grid::BlockGrid, idx::CartesianIndex)

Linear index in grid.remote_blocks of the remote block at idx in the grid.

Armon.block_pos_containing_cell — Function

block_pos_containing_cell(grid::BlockGrid, pos::Union{CartesianIndex, NTuple})

Returns two CartesianIndexes: the first is the position of the block containing the cell at pos, the second is the position of the cell in that block.

Armon.block_origin — Function

block_origin(grid::BlockGrid, pos, include_ghosts=false)

A Tuple of the position of the cell at the bottom left corner of the LocalTaskBlock at pos in the grid.

If include_ghosts == true, then the cell position includes all ghost cells of the grid.

pos can be any of Integer, NTuple{N, Integer} or CartesianIndex.

Armon.block_at — Function

block_at(grid::BlockGrid, idx::CartesianIndex)

The TaskBlock at position idx in the grid.

Armon.block_size_at — Function

block_size_at(grid::BlockGrid, idx)
block_size_at(idx, grid_size, static_sized_grid, block_size, remainder_block_size, ghosts)

Theoretical size of block at idx in grid. If ghosts == 0, ghosts cells will be included in the size.

Iteration

Armon.all_blocks — Function

all_blocks(grid::BlockGrid)

Simple iterator over all blocks of the grid, excluding RemoteTaskBlocks.

Armon.@iter_blocks — Macro

@iter_blocks for blk in grid
    # body...
end

Applies the body of the for-loop in to all blocks of the grid. Threads iterate over the blocks they are assigned to via grid.threads_workload.

Armon.EdgeBlockRegions — Type

EdgeBlockRegions(grid::BlockGrid; real_sizes=false)
EdgeBlockRegions(
    grid_size::NTuple{D, Int}, static_sized_grid::NTuple{D, Int},
    block_size::NTuple{D, Int}, remainder_block_size::NTuple{D, Int}; ghosts=0
)

Iterator over edge blocks, their positions and their size in each edge region of the grid.

ghosts only affects the size of edge blocks: B .- 2*ghosts. If real_sizes == true, then ghosts is ghosts(grid), therefore the real block size of edge blocks is returned.

There is a maximum of 2^D-1 edge regions in a grid (see this explanation). When grid_size[i] != static_sized_grid[i], there cannot be an edge region, hence there are exactly 2^sum(grid_size .!= static_sized_grid)-1 regions.

julia> collect(Armon.EdgeBlockRegions((5, 5), (4, 4), (32, 32), (16, 16)))
3-element Vector{Tuple{Int64, CartesianIndices{2, Tuple{UnitRange{Int64}, UnitRange{Int64}}}, Tuple{Int64, Int64}}}:
 (4, CartesianIndices((5:5, 1:4)), (16, 32))
 (4, CartesianIndices((1:4, 5:5)), (32, 16))
 (1, CartesianIndices((5:5, 5:5)), (16, 16))

julia> collect(Armon.EdgeBlockRegions((5, 5, 5), (4, 5, 4), (32, 32, 32), (16, 16, 16)))
3-element Vector{Tuple{Int64, CartesianIndices{3, Tuple{UnitRange{Int64}, UnitRange{Int64}, UnitRange{Int64}}}, Tuple{Int64, Int64, Int64}}}:
 (20, CartesianIndices((5:5, 1:5, 1:4)), (16, 32, 32))
 (20, CartesianIndices((1:4, 1:5, 5:5)), (32, 32, 16))
 (5, CartesianIndices((5:5, 1:5, 5:5)), (16, 32, 16))

Armon.RemoteBlockRegions — Type

RemoteBlockRegions(grid::BlockGrid)
RemoteBlockRegions(grid_size::NTuple{D, Int})

Iterator of all remote block positions in each region (the faces of the grid). There is always 2*D regions.

julia> collect(Armon.RemoteBlockRegions((5, 3)))
4-element Vector{Tuple{Armon.Side.T, CartesianIndices{2, Tuple{UnitRange{Int64}, UnitRange{Int64}}}}}:
 (Armon.Side.Left, CartesianIndices((0:0, 1:3)))
 (Armon.Side.Right, CartesianIndices((6:6, 1:3)))
 (Armon.Side.Bottom, CartesianIndices((1:5, 0:0)))
 (Armon.Side.Top, CartesianIndices((1:5, 4:4)))

Armon.BlockRowIterator — Type

BlockRowIterator(grid::BlockGrid; kwargs...)
BlockRowIterator(grid::BlockGrid, blk::LocalTaskBlock; kwargs...)
BlockRowIterator(grid::BlockGrid, blk₁_pos, blk₂_pos; kwargs...)
BlockRowIterator(grid::BlockGrid, sub_domain::NTuple{2, CartesianIndex}; kwargs...)
BlockRowIterator(grid::BlockGrid, row_iter::CartesianIndices; global_ghosts=false, all_ghosts=false)

Iterate the rows of all blocks of the grid, row by row (and not block by block). This allows to iterate the cells of the grid as if it was a single block.

Elements are tuples of (block, global_row_idx, row_range). row_range is the range of cells in block for the current row.

Giving blk will return an iterator on the rows of the block.

Giving blk₁_pos and blk₂_pos will return an iterator over all rows between those blocks.

Giving sub_domain will return an iterator including only the cells contained in sub_domain. sub_domain is a cuboid defined by the position of the cells in the whole domain of grid, using its lower and upper corners.

row_iter is a iterator over global row indices.

If global_ghosts == true, then the ghost cells of at the border of the global domain are also returned. If all_ghosts == true, then the ghost cells of at the border of all blocks are also returned.

julia> params = ArmonParameters(; N=(24, 8), nghost=4, block_size=(20, 12), use_MPI=false, numa_aware=false);

julia> grid = BlockGrid(params);

julia> for (blk, row_idx, row_range) in Armon.BlockRowIterator(grid)
           println(Tuple(blk.pos), " - ", row_idx, " - ", row_range)
       end
(1, 1) - (1, 1) - 85:96
(2, 1) - (2, 1) - 85:96
(1, 1) - (1, 2) - 105:116
(2, 1) - (2, 2) - 105:116
(1, 1) - (1, 3) - 125:136
(2, 1) - (2, 3) - 125:136
(1, 1) - (1, 4) - 145:156
(2, 1) - (2, 4) - 145:156
(1, 2) - (1, 5) - 85:96
(2, 2) - (2, 5) - 85:96
(1, 2) - (1, 6) - 105:116
(2, 2) - (2, 6) - 105:116
(1, 2) - (1, 7) - 125:136
(2, 2) - (2, 7) - 125:136
(1, 2) - (1, 8) - 145:156
(2, 2) - (2, 8) - 145:156

Host/Device transfers

Armon.device_to_host! — Function

device_to_host!(blk::LocalTaskBlock)

Copies the device data of blk to the host data. A no-op if the device is the host.

device_to_host!(grid::BlockGrid)

Copies device data of all blocks to the host data. A no-op if the device is the host.

Armon.host_to_device! — Function

device_to_host!(blk::LocalTaskBlock)

Copies the host data of blk to its device data. A no-op if the device is the host.

device_to_host!(grid::BlockGrid)

Copies host data of all blocks to the device data. A no-op if the device is the host.

Armon.buffers_on_device — Function

buffers_on_device(::BlockGrid)
buffers_on_device(::Type{<:BlockGrid})

true if the communication buffers are stored on the device, allowing direct transfers without passing through the host (GPU-aware communication).

Armon.device_is_host — Function

device_is_host(::BlockGrid{T, D, H})
device_is_host(::Type{<:BlockGrid{T, D, H}})

true if the device is the host, i.e. device blocks and host blocks are the same (and D == H).

NUMA

Armon.move_pages — Method

move_pages(grid::BlockGrid)

Move the pages of all blocks of the grid, including remote blocks, to the NUMA node of the thread which is in charge of working on that block.

Armon.lock_pages — Method

lock_pages(grid::BlockGrid)

Locks the pages of all blocks of the grid, including remote blocks.

Block distribution

Armon.thread_workload_distribution — Function

thread_workload_distribution(params::ArmonParameters; threads=nothing)
thread_workload_distribution(
    threads::Int, grid_size::Tuple;
    scotch=true, simple=false, perimeter_first=false, kwargs...
)

Distribute each block in grid_size among the threads, as evenly as possible.

With simple == true, blocks are distributed with simple_workload_distribution.

With scotch == true, the Scotch graph partitioning solver is used to better split the grid. kwargs are passed to scotch_grid_partition.

If perimeter_first == true, the resulting distribution will have blocks sorted in way that will place neighbours of other threads' blocks first in the list. By doing so, communications between threads may be overlapped more frequently.

Armon.simple_workload_distribution — Function

simple_workload_distribution(threads, grid_size)

Basic distribution of B = prod(grid_size) blocks into groups of B ÷ threads, with the remaining blocks distributed evenly.

Armon.scotch_grid_partition — Function

scotch_grid_partition(
    threads, grid_size;
    strategy=:default, workload_tolerance=0, repart=false, retries=10, weighted=false,
    static_sized_grid=nothing, block_size=nothing, remainder_block_size=nothing, ghosts=0
)

Split grid_size to the threads.

strategy is passed to Scotch.strat_flags.

workload_tolerance is the acceptable workload uneveness among the partitions. Giving a few blocks of margin (e.g. at least 1/prod(grid_size)) is preferrable.

weighted == true will distribute blocks while taking into account the number of real cells they have. In this case all parameters of the grid must be present: static_sized_grid, block_size, remainder_block_size and ghosts must be given (obtained with e.g. grid_dimensions).

The partitioning is random, hence results may vary. To counterbalance this, giving retries > 0 will repeat the partitioning retries times and keep the best one.

Armon.block_grid_from_workload — Function

block_grid_from_workload(grid_size, threads_workload)

Convenience function to convert a threads_workload (result of thread_workload_distribution) into an Array of grid_size, with each element assigned to the tid given by the distribution.

This makes it easy to visualize the efficiency of the distribution.

Logging

Armon.BlockLogEvent — Type

BlockLogEvent

Info about a block, emitted after a call to block_state_machine which successfully advanced the internal state of the block, only if params.log_blocks == true.

Armon.ThreadLogEvent — Type

ThreadLogEvent

Info about a thread, emitted after a call to solver_cycle_async, which itself calls block_state_machine.

Armon.collect_logs — Function

collect_logs(grid::BlockGrid)

Collect all BlockLogEvent of the grid into a single object (a BlockGridLog). See analyse_log_stats to print metrics about all blocks.

Armon.analyse_log_stats — Function

analyse_log_stats(f, grid_log::BlockGridLog)

Call f for each block of the grid_log, passing the block's position (a CartesianIndex), a Vector of BlockLogEvents, and the size of the block.

analyse_log_stats(grid_log::BlockGridLog)

Crunch all data of grid_log into tangible metrics contained in a BlockGridLogStats.

Armon.BlockGridLogStats — Type

BlockGridLogStats

Metrics about the blocks and threads of a full solver excution.

Armon.BLOCK_LOG_THREAD_LOCAL_STORAGE — Constant

BLOCK_LOG_THREAD_LOCAL_STORAGE::Dict{UInt16, Int32}

Incremented by 1 every time a BlockLogEvent is created in a thread, i.e. each time a block has solver kernels applied to it through block_state_machine.

Since only differences between values are interesting, no need to reset it.