Kernels

Armon.@generic_kernel — Macro

@generic_kernel(function definition)

Transforms a single kernel function into six different functions:

4 which run on the CPU using Polyester.jl's multi-threading or not, as well as SIMD or not.

one which uses KernelAbstractions.jl to make a GPU-compatible kernel

a main function, which will take care of calling the two others depending if we want to use the GPU or not.

To do this, two things are done:

All calls to @index_1D_lin(), @index_2D_lin() and @iter_idx() are replaced by their equivalent in their respective platforms: a simple loop index for CPUs, and a call to KA.jl's @index for GPUs.

Arguments to each function are edited

A kernel function must call one of @index_1D_lin() or @index_2D_lin() at least once, since this will determine which type of indexing to use as well as which parameters to add.

The indexing macro @iter_idx gives the linear index to the current iteration (on CPU) or global thread (on GPU).

This means that in order to call the new main function, one needs to take into account which indexing macro was used:

In all cases, params::ArmonParameters is the first argument

Then, depending on the indexing macro used:

@index_1D_lin() : loop_range::OrdinalRange{Int}

@index_2D_lin() : main_range::OrdinalRange{Int}, inner_range::OrdinalRange{Int}

An optional keyword argument, no_threading, allows to override the use_threading parameter, which can be useful in asynchronous contexts. It defaults to false.

Using KA.jl's @Const to annotate arguments is supported, but they will be present only in the GPU kernel definition.

Further customisation of the kernel and main function can be obtained using @kernel_options and @kernel_init.

@generic_kernel f_kernel(A)
    i = @index_1D_lin()
    A[i] += 1
end

params.use_gpu = false
f(params, 1:10)  # CPU call

params.use_gpu = true
f(params, 1:10)  # GPU call

Armon.@kernel_init — Macro

@kernel_init(expr)

Allows to initialize some internal variables of the kernel before the loop. The given expression must NOT depend on any index. You must not use any indexing macro (@index_1D_lin(), etc...) in the expression.

All paramerters of the kernel are available during the execution of the init expression. On GPU, this expression will be executed by each thread.

This is a workaround for a limitation of Polyester.jl which prevents you from typing variables.

@generic_kernel function simple_kernel(a, b)
    @kernel_init begin
        c::Float64 = sin(0.123)
    end
    i = @index_1D_lin()
    a[i] += b[i] * c
end

Armon.@kernel_options — Macro

@kernel_options(options...)

Must be used (once, and explicitly) in the definition of a @generic_kernel function.

Gives options for @generic_kernel to adjust the resulting functions.

The possible options are:

debug: Prints the generated functions to stdout at compile time.

@generic_kernel function add_kernel(a, b)
    @kernel_options(debug)
    i = @index_1D_lin()
    a[i] += b[i]
end

Armon.@index_1D_lin — Macro

@index_1D_lin()

Indexing macro to use in a @generic_kernel function. Returns a linear index to access the 1D arrays used by the kernel.

Cannot be used alongside @index_2D_lin().

Armon.@index_2D_lin — Macro

@index_2D_lin()

Indexing macro to use in a @generic_kernel function. Returns a linear index to access the 2D arrays used by the kernel.

Cannot be used alongside @index_1D_lin().

Armon.@iter_idx — Macro

@iter_idx()

Indexing macro to use in a @generic_kernel function. Returns a linear index to access the 2D arrays used by the kernel.

Equivalent to KernelAbstractions.jl's @index(Global, Linear).

Armon.@simd_loop — Macro

@simd_loop(expr)

Allows to enable/disable SIMD optimisations for a loop. When SIMD is enabled, it is assumed that there is no dependencies between each iterations of the loop.

    @simd_loop for i = 1:n
        y[i] = x[i] * (x[i-1])
    end

Armon.@simd_threaded_iter — Macro

@simd_threaded_iter(range, expr)

Same as @simd_threaded_loop(expr), but instead of slicing the range of the for loop in expr, we slice the range given as the first parameter and distribute the slices evenly to the threads.

The inner @simd loop assumes there is no dependencies between each iteration.

    @simd_threaded_iter 4:2:100 for i in 1:100
        y[i] = log10(x[i]) + x[i]
    end
    # is equivalent to (without threading and SIMD)
    for j in 4:2:100
        for i in (1:100) .+ (j - 1)
            y[i] = log10(x[i]) + x[i]
        end
    end

Armon.@simd_threaded_loop — Macro

@simd_threaded_loop(expr)

Allows to enable/disable multithreading and/or SIMD of the loop depending on the parameters. When using SIMD, @fastmath and @inbounds are used.

In order to use SIMD and multithreading at the same time, the range of the loop is split in even batches. Each batch has a size of params.simd_batch iterations, meaning that the inner @simd loop has a fixed number of iterations, while the outer threaded loop will have N ÷ params.simd_batch iterations.

The loop range is assumed to be increasing, i.e. this is correct: 1:2:100, this is not: 100:-2:1 The inner @simd loop assumes there is no dependencies between each iteration.

    @simd_threaded_loop for i = 1:n
        y[i] = log10(x[i]) + x[i]
    end

Armon.@threaded — Macro

@threaded(expr)

Allows to enable/disable multithreading of the loop depending on the parameters.

The default condition is params.use_threading && !params.use_cache_blocking. By passing :outside_kernel before expr, the condition becomes simply params.use_threading.

    @threaded for i = 1:n
        y[i] = log10(x[i]) + x[i]
    end

Armon.@threads — Macro

Controls which multi-threading library to use.