Parameters and entry point

Armon.ArmonParameters — Type

ArmonParameters(; options...)

The parameters and current state of the solver.

The state is reset at each call to armon.

There are many options. Each backend can add their own.

Options

Backend and MPI

device = :CUDA

Device to use. Supported values:

:CPU_HP: Polyester.jl CPU multithreading (default if use_gpu=false)

:CUDA: CUDA.jl GPU (default if use_gpu=true)

:ROCM: AMDGPU.jl GPU

:CPU: KernelAbstractions.jl CPU multithreading (using the standard Threads.jl)

use_MPI = true, P = (1, 1), reorder_grid = true, global_comm = nothing

MPI config. The MPI domain will be a process grid of size P. global_comm is the global communicator to use, defaults to MPI.COMM_WORLD. reorder_grid is passed to MPI.Cart_create.

gpu_aware = true

Store MPI buffers on the device. This requires to use a GPU-aware MPI implementation. Does nothing when using the CPU only.

Kernels

use_threading = true, use_simd = true

Switches for CPU_HP kernels. use_threading enables @threaded for outer loops. use_simd enables @simd_loop for inner loops.

use_gpu = false

Enables the use of KernelAbstractions.jl kernels.

use_kokkos = false

Use kernels for Kokkos.jl.

use_cache_blocking = true

Separate the domain into semi-independant blocks, improving the cache-locality of memory accesses and therefore memory throughput.

async_cycle = false

Apply all steps of the solver to all blocks asynchronously, fully taking advantage of cache blocking.

block_size = 1024

Size of blocks for cache blocking. Can be a tuple. If use_cache_blocking == false, this option only controls the size of GPU blocks.

use_two_step_reduction = false

Reduction kernels (dtCFL_kernel and conservation_vars) use some optimizations to perform the reduction in a single step. It might cause issues on some GPU backends: a more "gentle" approach could avoid those by doing it in two steps.

Profiling

profiling = Symbol[]

List of profiling callbacks to use:

:TimerOutputs: TimerOutputs.jl sections (added if measure_time=true)

:NVTX_sections: NVTX.jl sections

:NVTX_kernels: NVTX.jl sections for kernels

:CUDA_kernels: equivalent to CUDA.@profile in front of all kernels

measure_time = true

measure_time=false can remove any overhead caused by profiling.

time_async = true

time_async=false will add a barrier at the end of every section. Useful for GPU kernels.

Scheme and CFD solver

scheme = :GAD, riemann_limiter = :minmod

scheme is the Riemann solver scheme to use:

:Godunov (1st order)

:GAD (2nd order, with limiter).

riemann_limiter is the limiter to use for the Riemann solver: :no_limiter, :minmod or :superbee.

projection = :euler_2nd

Scheme for the Eulerian remap step:

:euler (1st order)

:euler_2nd (2nd order, +minmod limiter)

axis_splitting = :Sequential

Axis splitting to use:

:Sequential: X then Y

:SequentialSym (or :Godunov): X and Y then Y and X, alternating

:Strang: ½X, Y, ½X then ½Y, X, ½Y, alternating (½ is for halved time step)

:X_only

:Y_only

N = (10, 10)

Number of cells of the global domain in each axes.

nghost = 4

Number of ghost cells. Must be greater or equal to the minimum number of ghost cells (min 1, scheme=:GAD adds one, projection=:euler_2nd adds one, scheme=:GAD + projection=:euler_2nd adds another one)

Dt = 0., cst_dt = false, dt_on_even_cycles = false

Dt is the initial time step, it is computed after initialization by default. If cst_dt=true then the time step is always Dt and no reduction over the entire domain occurs. If dt_on_even_cycles=true then then time step is only updated at even cycles (the first cycle is even).

data_type = Float64

Data type for all variables. Should be an AbstractFloat.

Test case and domain

test = :Sod, domain_size = nothing, origin = nothing

test is the test case name to use:

:Sod: Sod shock tube test

:Sod_y: Sod shock tube test along the Y axis

:Sod_circ: Circular Sod shock tube test (centered in the domain)

:Bizarrium: Bizarrium test, similar to the Sod shock tube but with a special equation of state

:Sedov: Sedov blast-wave test (centered in the domain, reaches the border at t=1 by default)

:DebugIndexes: Set all variables to their index in the global domain. Debug only.

cfl = 0., maxtime = 0., maxcycle = 500_000

cfl defaults to the test's default value, same for maxtime. The solver stops when t reaches maxtime or maxcycle iterations were done (maxcycle=0 stops after initialization).

Output

silent = 0

silent=0 for maximum verbosity. silent=3 doesn't print info at each cycle. silent=5 doesn't print anything.

output_dir = ".", output_file = "output"

joinpath(output_dir, output_file) will be path to the output file.

write_output = false, write_ghosts = false

write_output=true will write all saved_vars() to the output file. If write_ghosts=true, ghost cells will also be included.

write_slices = false

Will write all saved_vars() to 3 output files, one for the middle X row, another for the middle Y column, and another for the diagonal. If write_ghosts=true, ghost cells will also be included.

output_precision = nothing

Numbers are saved with output_precision digits of precision. Defaults to enough numbers for an exact decimal representation.

animation_step = 0

If animation_step ≥ 1, then every animation_step cycles, variables will be saved as with write_output=true.

compare = false, is_ref = false, comparison_tolerance = 1e-10

If compare=true, then at every sub step of each iteration of the solver all variables will:

(is_ref=false) be compared with a reference file found in output_dir

(is_ref=true) be saved to a reference file in output_dir

When comparing, a relative comparison_tolerance (the rtol kwarg of isapprox) is accepted between values.

check_result = false

Check if conservation of mass and energy is verified between initialization and the last iteration. An error is thrown otherwise. Accepts a relative comparison_tolerance.

return_data = false

If return_data=true, then in the SolverStats returned by armon, the data field will contain the BlockGrid used by the solver.

Armon.armon — Function

armon(::ArmonParameters)

Main entry point of the solver. Returns a SolverStats.

Armon.SolverStats — Type

SolverStats

Solver output.

data is nothing if parameters.return_data is false.

timer is nothing if parameters.measure_time is false.

Armon.StepsRanges — Type

StepsRanges

Holds indexing information for all steps of the solver.

Domains are stored as block corner offsets: blocks can have different sizes, but always the same amount of ghost cells, therefore the iteration domain is determined from the dimensions of the block. The first field is the offset to the first cell, the second is the offset to the last cell.

Armon.data_type — Function

data_type(::ArmonParameters{T})

Get T, the type used for numbers by the solver

Device and backends

Armon.CPU_HP — Type

CPU_HP

Device tag for the high-performance CPU backend using multithreading (with Polyester.jl) and vectorisation.

Armon.create_device — Function

create_device(::Val{:device_name})

Create a device object from its name.

Default devices:

:CPU: the CPU backend of KernelAbstractions.jl

:CPU_HP: Polyester.jl multithreading

Extensions:

:Kokkos: the default Kokkos.jl device

:CUDA: the CUDA.jl backend of KernelAbstractions.jl

:ROCM: the AMDGPU.jl backend of KernelAbstractions.jl

Armon.init_backend — Function

init_backend(params::ArmonParameters, ::Dev; options...)

Initialize the backend corresponding to the Dev device returned by create_device using options. Set the params.backend_options field.

It must return options, with the backend-specific options removed.

Armon.device_memory_info — Function

device_memory_info(device)

The total and free memory on the device, in bytes.

Armon.memory_info — Function

memory_info(params)

The total and free memory the current process can store on the params.device.

Armon.memory_required — Function

memory_required(params::ArmonParameters)

(device_memory, host_memory) required for params.

MPI buffers size are included in the appropriate field depending on params.gpu_aware. params.use_MPI and params.neighbours is taken into account.

If device_is_host, then, device_memory only includes memory required by data arrays and MPI buffers.

memory_required(N::NTuple{2, Int}, block_size::NTuple{2, Int}, ghost::Int, data_type)
memory_required(N::NTuple{2, Int}, block_size::NTuple{2, Int}, ghost::Int,
    device_array_type, host_array_type, buffer_array_type[, solver_state_type])

Compute the number of bytes needed to allocate all blocks. If only data_type is given, then device_array_type, host_array_type and buffer_array_type default to Vector{T}. solver_state_type defaults to SolverState{T, #= default schemes and test =#}.

In order of returned values:

Amount of bytes needed for all arrays on the device. This amount is also required on the host when the host and device are not the same.

Amount of bytes needed for all MPI buffers, if the sub-domain has neighbours on all of its sides. If params.gpu_aware, then this memory is allocated on the device.

Amount of bytes needed on the host memory for all block objects, excluding array data and buffers. This memory is always allocated on the host.

res = memory_required((1000, 1000), (64, 64), 4, CuArray{Float64}, Vector{Float64}, Vector{Float64})
device_memory = res[1]
host_memory = res[3] + (device_is_host ? device_memory : res[1])
if params.gpu_aware
    device_memory += res[2]
else
    host_memory += res[2]
end