Utilities

Armon.Axis — Module

Axis

Enumeration of the axes of the domain

Armon.Side — Module

Side

Enumeration of the sides of the domain

Armon.SolverException — Type

SolverException

Thrown when the solver encounters an invalid state.

The category field can be used to distinguish between error types without inspecting the error message:

:config: a problem in the solver configuration, usually thrown when constructing ArmonParameters

:cpp: a C++ exception thrown by the C++ Kokkos backend

:time: an invalid time step

ErrorExceptions thrown by the solver represent internal errors.

Armon.@section — Macro

@section(name, expr)
@section(name, options, expr)

Introduce a profiling section around expr. Sections can be nested. Sections do not introduce a new scope.

Placed before a for-loop, a new section will be started for each iteration. name can interpolate using loop variables (like Test.@testset).

It is assumed that a params variable of ArmonParameters is present in the scope of the @section.

options is of the form key=value:

async (default: false): if async=false (and !params.time_async), a barrier (wait(params)) is added at the end of the section.

params = ArmonParameters(#= ... =#)

@section "Iteration $i" for i in 1:10
    j = @section "Foo" begin
        foo(i)
    end

    @section "Some calculation" begin
        k = bar(i, j)
    end

    @sync begin
        @async begin
            @section "Task 1" async=true my_task_1(i, j, k)
        end

        @async begin
            @section "Task 2" async=true my_task_2(i, j, k)
        end
    end
end

NUMA utilities

Armon.array_pages — Function

array_pages(A::Ptr, A_length)
array_pages(A::DenseArray)

Iterator over the memory pages (aligned to PAGE_SIZE) used by the array A.

Armon.touch_pages — Function

touch_pages(pages)

Parses over each of pages and "touch" them by loading a value then storing it again.

The (only) use case is to manually trigger the first-touch policy and/or force the kernel to physically allocate all pages.

Armon.move_pages — Method

move_pages(pages::Vector{Ptr}, target_node; retries=4)
move_pages(pages::OrdinalRange{Ptr, Int}, target_node)
move_pages(A::DenseArray, target_node)

Move all pages (of either the array A or from the pages iterable) to the target_node (1-index).

The move status of all pages is checked afterwards. No error is thrown only if all pages are on target_node at the end of the move.

In case some pages cannot be moved because of EBUSY status, the move is retried retries-times until it succeeds, with a short sleep in-between attempts.

Armon.lock_pages — Method

lock_pages(ptr, len)
lock_pages(pages::OrdinalRange{Ptr, Int})
lock_pages(A::DenseArray)

Lock the pages ranging from ptr to ptr+len on the RAM, preventing the kernel from moving it around or putting it in swap memory.

Armon.unlock_pages — Function

unlock_pages(ptr, len)
unlock_pages(pages::OrdinalRange{Ptr, Int})
unlock_pages(A::DenseArray)

Unlock the pages, opposite of lock_pages.