Utility

Armon.Axis — Module

Axis

Enumeration of the axes of the domain

Armon.Side — Module

Side

Enumeration of the sides of the domain

Armon.SolverException — Type

SolverException

Thrown when the solver encounters an invalid state.

The category field can be used to distinguish between error types without inspecting the error message:

:config: a problem in the solver configuration, usually thrown when constructing ArmonParameters

:cpp: a C++ exception thrown by the C++ Kokkos backend

:time: an invalid time step

ErrorExceptions thrown by the solver represent internal errors.

Armon.threads_workload — Function

threads_workload(params::ArmonParameters, grid)

Estimate the workload of each thread when using solver_cycle_async (i.e. params.async_cycle == true). Returns a thread_count × 3 Matrix: the first column is the thread ID (sorted by ascending order), the second is the number of batches, the third is the number of LocalTaskBlock.

Armon.@section — Macro

@section(name, expr)
@section(name, options, expr)

Introduce a profiling section around expr. Sections can be nested. Sections do not introduce a new scope.

Placed before a for-loop, a new section will be started for each iteration. name can interpolate using loop variables (like Test.@testset).

It is assumed that a params variable of ArmonParameters is present in the scope of the @section.

options is of the form key=value:

async (default: false): if async=false (and !params.time_async), a barrier (wait(params)) is added at the end of the section.

params = ArmonParameters(#= ... =#)

@section "Iteration $i" for i in 1:10
    j = @section "Foo" begin
        foo(i)
    end

    @section "Some calculation" begin
        k = bar(i, j)
    end

    @sync begin
        @async begin
            @section "Task 1" async=true my_task_1(i, j, k)
        end

        @async begin
            @section "Task 2" async=true my_task_2(i, j, k)
        end
    end
end